services/gRPC/requirements.txt
grpcio
grpcio-tools
protobuf
requests
runtime